Moved dataset files, added sample file for housing data.
