+++ tests/dataSet.test.js
@@ -4,11 +4,9 @@ require('url-search-params-polyfill');
 
 const chai = require('chai');
 const expect = chai.expect;
-const testDataSetDetail = require('./dataGenerator');
-
+const testDataSetDetail = require('./fixtures/time-series.json');
 
 import DataSetClient from '../src/DataSetClient';
-
 
 describe('DataSet tests', () => {
     let client = new DataSetClient({ endpoint: global.endpointUrl, key: process.env.NEXOSIS_API_TESTKEY });
